refactor : 사용하지 않는 WinnerNumbers 도메인 제거 및 Lotto 패키지 이동
