+++ src/main/java/lotto/service/LottoShop.java
@@ -2,7 +2,7 @@ package lotto.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.LottoNumberGenerator;
 
 public class LottoShop {
